+++ keyboards/orthopanda/MCP23S17/MCP23S17.h
@@ -38,26 +38,26 @@
 // Variables
 // extern uint8_t MCP_address;                        // Address of the MCP23S17 in use
 // extern pin_t MCP_slaveSelectPin;                   // Slave-select pin
-// extern unsigned int MCP_modeCache;                 // Caches the mode (input/output) configuration of I/O pins
-// extern unsigned int MCP_pullupCache;               // Caches the internal pull-up configuration of input pins (values persist across mode changes)
-// extern unsigned int MCP_invertCache;               // Caches the input pin inversion selection (values persist across mode changes)
-// extern unsigned int MCP_outputCache;               // Caches the output pin state of pins
+// extern uint16_t MCP_modeCache;                 // Caches the mode (input/output) configuration of I/O pins
+// extern uint16_t MCP_pullupCache;               // Caches the internal pull-up configuration of input pins (values persist across mode changes)
+// extern uint16_t MCP_invertCache;               // Caches the input pin inversion selection (values persist across mode changes)
+// extern uint16_t MCP_outputCache;               // Caches the output pin state of pins
 
 // Functions
 void MCP_init(uint8_t address, pin_t ss);            // Constructor to instantiate a discrete IC as an object, address 0-7, chipSelect any valid pin
 bool MCP_begin(void);                                // Start the SPI Bus
-bool MCP_wordWrite(uint8_t reg, unsigned int word);  // Typically only used internally, but allows the user to write any register pair if needed, so it's public
+bool MCP_wordWrite(uint8_t reg, uint16_t word);  // Typically only used internally, but allows the user to write any register pair if needed, so it's public
 bool MCP_byteWrite(uint8_t reg, uint8_t value);      // Typically only used internally, but allows the user to write any register if needed, so it's public
 bool MCP_pinMode(uint8_t pin, uint8_t mode);         // Sets the mode (input or output) of a single I/O pin
-bool MCP_pinModeAll(unsigned int mode);              // Sets the mode (input or output) of all I/O pins at once
+bool MCP_pinModeAll(uint16_t mode);              // Sets the mode (input or output) of all I/O pins at once
 bool MCP_pullupMode(uint8_t pin, uint8_t mode);      // Selects internal 100k input pull-up of a single I/O pin
-bool MCP_pullupModeAll(unsigned int mode);           // Selects internal 100k input pull-up of all I/O pins at once
+bool MCP_pullupModeAll(uint16_t mode);           // Selects internal 100k input pull-up of all I/O pins at once
 bool MCP_inputInvert(uint8_t pin, uint8_t mode);     // Selects input state inversion of a single I/O pin (writing 1 turns on inversion)
-bool MCP_inputInvertAll(unsigned int mode);          // Selects input state inversion of all I/O pins at once (writing a 1 turns on inversion)
+bool MCP_inputInvertAll(uint16_t mode);          // Selects input state inversion of all I/O pins at once (writing a 1 turns on inversion)
 bool MCP_digitalWrite(uint8_t pin, uint8_t value);   // Sets an individual output pin HIGH or LOW
-bool MCP_digitalWriteAll(unsigned int value);        // Sets all output pins at once. If some pins are configured as input, those bits will be ignored on write
+bool MCP_digitalWriteAll(uint16_t value);        // Sets all output pins at once. If some pins are configured as input, those bits will be ignored on write
 uint8_t MCP_byteRead(uint8_t reg);                   // Reads an individual register and returns the byte. Argument is the register address
 uint8_t MCP_digitalRead(uint8_t pin);                // Reads an individual input pin
-unsigned int MCP_digitalReadAll(void);               // Reads all input  pins at once. Be sure it ignore the value of pins configured as output!
+uint16_t MCP_digitalReadAll(void);               // Reads all input  pins at once. Be sure it ignore the value of pins configured as output!
 
 #endif // MCP23S17
